--- a/crates/uv/src/bin/uv.rs
+++ b/crates/uv/src/bin/uv.rs
@@ -1,3 +1,7 @@
+// Don't optimize the alloc crate away due to it being otherwise unused.
+// https://github.com/rust-lang/rust/issues/64402
+extern crate uv_performance_memory_allocator;
+
 use std::process::ExitCode;
 
 use uv::main as uv_main;