@charliermarsh 2025-01-02
I'm running a dry-run release: https://github.com/astral-sh/uv/actions/runs/12584094167